Fix Linux surface retrieval by casting X11 window to ULong
The X11 window handle is now explicitly cast to ULong to ensure compatibility with the Vulkan surface creation. This resolves potential type mismatches and prevents errors during runtime on Linux systems.

+++ demo/desktop-and-ios/src/jvmMain/kotlin/main.kt
@@ -65,7 +65,7 @@ private fun getSurface(instance: WGPUInstance, window: Long): WGPUSurface = when
     Os.Linux -> {
         val display = glfwGetX11Display().toNativeAddress()
         val x11_window = glfwGetX11Window(window)
-        getSurfaceFromX11Window(instance, display, x11_window) ?: error("fail to get surface on Linux")
+        getSurfaceFromX11Window(instance, display, x11_window.toULong()) ?: error("fail to get surface on Linux")
     }
     Os.Window -> {
         val hwnd = glfwGetWin32Window(window).toNativeAddress()
